% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log.R
\name{archive_log}
\alias{archive_log}
\title{Archive logs}
\usage{
archive_log(regex = ".*", log_dir = "logs", new_dir = NULL)
}
\arguments{
\item{regex}{Regular expression a log must match to be archived}

\item{log_dir}{The logging directory.}

\item{new_dir}{The directory where the logs should be moved to. `NULL`
indicates a time stamped directory under "logs".}
}
\description{
Archive logs
}
