% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log.R
\name{archive_logs}
\alias{archive_logs}
\title{Archive logs}
\usage{
archive_logs(regex = ".*", log_dir = "logs", new_dir = NULL,
  create_empty = FALSE)
}
\arguments{
\item{regex}{Regular expression a log must match to be archived}

\item{log_dir}{The logging directory.}

\item{new_dir}{The directory where the logs should be moved to. `NULL`
indicates a time stamped directory under "logs".}

\item{create_empty}{Whether or not to create an empty directory.}
}
\description{
Archive logs
}
