% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manage.R
\name{delete_logs}
\alias{delete_logs}
\title{Delete logs with a checkpoint older than a reference date}
\usage{
delete_logs(relieve_date = Sys.Date() - 180, communicator = flog_info)
}
\arguments{
\item{communicator}{The bare name of a function that communicates the
how many logs are deleted.}

\item{relieve}{date The date that marks the cut-off for discarding logs.
All logs older than that will be deleted.}
}
\description{
Delete logs with a checkpoint older than a reference date
}
