% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base.R
\docType{package}
\name{drogger-package}
\alias{drogger}
\alias{drogger-package}
\title{A logging scheme for drake}
\description{
Utility functions to log drake build processes. They work as follows:
}
\details{
* put flog_start() before drake::make()
* put flog_stop() after drake::make()

You will get two types of log stored into ./logs:

* A log with named with a time stamp for each run you executed.
* A log called `all.txt` that contains the whole log up to now, in reversed
  order, i.e. the top line corresponds to the last logging entry, the last
  row to the first logging entry.

You can open the last log with `open_log()`.
You can find the last log file with `find_last_log()`.
}
\author{
\strong{Maintainer}: Lorenz Walthert \email{lorenz.walthert@icloud.com}

}
