% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base.R
\name{flog_init}
\alias{flog_init}
\title{Initialize a logger}
\usage{
flog_init(file_name = file.path("logs", "all.txt"),
  name = background_file_logger())
}
\arguments{
\item{file_name}{The name fo the file where the log should be saved to}

\item{name}{The name of the logger.}
}
\description{
Initialize a logger
}
