% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base.R
\name{flog_stop}
\alias{flog_stop}
\title{Stop Logging}
\usage{
flog_stop(name = background_file_logger(),
  msg = "Completed logging run")
}
\arguments{
\item{name}{The name of the logger.}

\item{msg}{An initial message to pass to the logger.}
}
\value{
Invisibly returns the path to the file that contains the log.
}
\description{
Stop Logging
}
