% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base.R
\name{flog_template}
\alias{flog_template}
\alias{flog_trace}
\alias{flog_info}
\alias{flog_debug}
\alias{flog_warn}
\alias{flog_error}
\alias{flog_fatal}
\title{Loggers}
\usage{
flog_template(logger, ..., name = background_file_logger())

flog_trace(...)

flog_info(...)

flog_debug(...)

flog_warn(...)

flog_error(...)

flog_fatal(...)
}
\arguments{
\item{logger}{The loging function to use.}

\item{...}{Message passed to the logger via `c(...)`.}

\item{name}{The name of the logger to use.}
}
\description{
These are wrappers around [futile.logger::flog_error()] and friends. They
write to the background logger by default and use `...` for message creation.
}
