% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base.R
\name{flog_threshold}
\alias{flog_threshold}
\title{Define the threshold for logging}
\usage{
flog_threshold(treshold, name)
}
\arguments{
\item{name}{character The name of the logger}
}
\description{
Wrapper around [flog.threshold()].
}
