% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{parse_log}
\alias{parse_log}
\title{Parse a log file}
\usage{
parse_log(path)
}
\arguments{
\item{path}{The path to a file that is a log.}
}
\description{
Parses a log that follows the structure used by `futile.loggeer`, i.e.
`INFO [2018-09-11 23:59:29] Here is a message.`
}
