% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base.R
\name{flog_start}
\alias{flog_start}
\title{Start logging}
\usage{
flog_start(threshold = futile.logger::INFO,
  name = background_file_logger(), msg = "Started logging run")
}
\arguments{
\item{threshold}{The new threshold for the given logger.}

\item{name}{The name of the logger.}

\item{msg}{An initial message to pass to the logger.}
}
\value{
Invisibly returns the path to the file that contains the log.
}
\description{
Start logging
}
